+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/dto/restaurantreview/RestaurantReviewRequestDto.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.dto.restaurant;
+package ProjectDoge.StudentSoup.dto.restaurantreview;
 
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
